Refs #38. Checkpointing refactoring.
