+++ src/interface/gui/VdrivePeakPicker.ui
@@ -789,6 +789,16 @@
         </item>
         <item>
          <layout class="QGridLayout" name="gridLayout_2">
+          <item row="0" column="2">
+           <widget class="QPushButton" name="pushButton_peakPickerMode">
+            <property name="toolTip">
+             <string>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;Enter or leave peak picking mode&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
+            </property>
+            <property name="text">
+             <string>Pick Mode</string>
+            </property>
+           </widget>
+          </item>
           <item row="0" column="1">
            <widget class="QPushButton" name="pushButton_showPeaks">
             <property name="toolTip">
@@ -833,16 +843,6 @@
             </property>
            </widget>
           </item>
-          <item row="0" column="0">
-           <widget class="QPushButton" name="pushButton_findPeaks">
-            <property name="toolTip">
-             <string>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;Find all peaks in the diffraction pattern&lt;/p&gt;&lt;p&gt;Case 1: If user does not provide structure information, then it will run FindPeaks();&lt;/p&gt;&lt;p&gt;Case 2: If user gives known structure, then this one will just plot peaks according to lattice parameters.&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
-            </property>
-            <property name="text">
-             <string>Find Peaks</string>
-            </property>
-           </widget>
-          </item>
           <item row="0" column="6">
            <spacer name="horizontalSpacer">
             <property name="orientation">
@@ -858,6 +858,16 @@
              </size>
             </property>
            </spacer>
+          </item>
+          <item row="0" column="0">
+           <widget class="QPushButton" name="pushButton_findPeaks">
+            <property name="toolTip">
+             <string>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;Find all peaks in the diffraction pattern&lt;/p&gt;&lt;p&gt;Case 1: If user does not provide structure information, then it will run FindPeaks();&lt;/p&gt;&lt;p&gt;Case 2: If user gives known structure, then this one will just plot peaks according to lattice parameters.&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
+            </property>
+            <property name="text">
+             <string>Find Peaks</string>
+            </property>
+           </widget>
           </item>
           <item row="0" column="4">
            <widget class="QPushButton" name="pushButton_deletePeaks">
@@ -893,16 +903,6 @@
             </property>
             <property name="text">
              <string>Set Overlap</string>
-            </property>
-           </widget>
-          </item>
-          <item row="0" column="2">
-           <widget class="QPushButton" name="pushButton_peakPickerMode">
-            <property name="toolTip">
-             <string>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;Enter or leave peak picking mode&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
-            </property>
-            <property name="text">
-             <string>Pick Mode</string>
             </property>
            </widget>
           </item>
